pyproject.toml
[project]
name = "mcp-multiserver-client-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.25",
    "langchain-experimental>=0.3.4",
    "langchain-groq>=0.3.2",
    "langchain-mcp-adapters>=0.1.7",
    "langchain-openai>=0.3.24",
    "langgraph>=0.4.8",
    "mcp[cli]==1.9.4",
    "openai>=1.88.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "streamlit>=1.46.0",
]
